Robust builder test
